secure_update: Include initial metadata with pip
We need to ship initial metadata for supported repositories (read
pypi.org) with pip. This is used to populate the actual metadata
directory if it does not exist.

This commit includes a placeholder metadata for repository
"https://localhost:8000/simple/": it should be replaced with pypi.org
metadata once it is available.